installer WIP

+++ lutris/installer.py
@@ -1,60 +1,119 @@
 import os
+import logging
 import urllib
+import urllib2
 import subprocess
 import shutil
+import yaml
 import platform
 from lutris.config import LutrisConfig
+import lutris.constants
+from distutils.command.install_data import install_data
 
 class Installer():
-    def __init__(self, gameName):
+    def __init__(self, game):
+        self.supported_protocol_version = 1
+
         self.lutris_config = LutrisConfig()
-        self.gameName = gameName
-        self.gamesDir = "/mnt/seagate/games/"
-        self.lutrisConfigDir = self.lutris_config.lutris_config_path
-        self.totalDownloaded = 0
+        self.website_url = lutris.constants.installer_prefix
+        self.installer_cache = lutris.constants.cache_path
+        self.gameName = game
+        self.game_info = {}
+        #FIXME : The thing that should not be
+        self.gamesDir = "/media/seagate300/games/"
         self.gameDir = os.path.join(self.gamesDir, self.gameName)
         if not os.path.exists(self.gameDir):
             os.mkdir(self.gameDir)
-        self.parseconfig()
-        self.writeConfig()
+        self.lutrisConfigDir = lutris.constants.lutris_config_path
+        self.installer_dest_path = os.path.join(self.installer_cache, self.gameName + ".yml")
+        #WTF ?
+        self.totalDownloaded = 0
+
+    def install(self):
+        self.save_installer_content()
+        success = self.parseconfig()
+        if success:
+            self.writeConfig()
+        else:
+            print "Installer failed"
+
+    def save_installer_content(self):
+        print 'downloading installer for ' + self.gameName
+        full_url = self.website_url + self.gameName + '.yml'
+        request = urllib2.Request(url = full_url)
+        f = urllib2.urlopen(request)
+
+        installer_file = open(self.installer_dest_path, "w")
+        installer_file.write(f.read())
+        installer_file.close()
+        #TODO : Check size
 
     def parseconfig(self):
-        installFile = open(self.gameName + ".lutris", "r")
-        self.installData = installFile.readlines()
-        for line in self.installData:
-            if line.startswith("execlink=") or line.startswith("datalink") or line.startswith("sourcelink"):
-                url = line[str.index(line, "=") + 1:].strip()
-                if url == "request":
-                    print "Please add the files manually"
-                else:
-                    destFile = self.download(url)
-                    os.chdir(self.gameDir)
-                    extIndex = destFile.rfind(".")
-                    extension = destFile[extIndex:]
-                    if extension == ".zip":
-                        self.unzip(destFile)
-                    elif extension == ".rar":
-                        self.unrar(destFile)
-                    elif extension == ".tgz" or extension == ".gz":
-                        self.untgz(destFile)
-                    elif extension == ".run" or extension == ".bin":
-                        self.runInstaller(destFile)
-            if line.startswith("exec_x86") and platform.architecture()[0] == "32bit":
-                self.exe = line[str.index(line, "=") + 1:].strip()
-            if line.startswith("exec_x64") and platform.architecture()[0] == "64bit":
-                self.exe = line[str.index(line, "=") + 1:].strip()
-            if line.startswith("platform"):
-                self.platform = line[str.index(line, "=") + 1:]
-        installFile.close()
+        self.install_data = yaml.load(file(self.installer_dest_path, 'r').read())
+
+        #Checking protocol
+        protocol_version = self.install_data['protocol']
+        if protocol_version != self.supported_protocol_version:
+            print "Wrong protocol version (Expected %d, got %d)" % (self.supported_protocol_version, protocol_version)
+            return False
+
+        #Script version
+        self.game_info['version'] = self.install_data['version']
+        #Runner
+        self.game_info['runner'] = self.install_data['runner']
+        #Name
+        self.game_info['name'] = self.install_data['name']
+        files = self.install_data['files']
+        self.gamefiles = {}
+        for gamefile in files:
+            file_id = gamefile.keys()[0]
+            dest_path = self.download(gamefile[file_id])
+            self.gamefiles[file_id] = dest_path
+        installer_actions = self.install_data['installer']
+
+        #FIXME ?
+        os.chdir(self.gameDir)
+
+        for action in installer_actions:
+            action_name = action.keys()[0]
+            action_data = action[action_name]
+            mappings = {
+                'check_md5': self.check_md5,
+                'extract' : self.extract,
+                'move' : self.move
+            }
+            if action_name not in mappings.keys():
+                print "Action " + action_name + " not supported !"
+                return False
+            mappings[action_name](action_data)
+#                    
+#                    extIndex = 
+#                    extension = destFile[destFile.rfind("."):]
+#                    if extension == ".zip":
+#                        self.unzip(destFile)
+#                    elif extension == ".rar":
+#                        self.unrar(destFile)
+#                    elif extension == ".tgz" or extension == ".gz":
+#                        self.untgz(destFile)
+#                    elif extension == ".run" or extension == ".bin":
+#                        self.runInstaller(destFile)
+        #=======================================================================
+        # if line.startswith("exec_x86") and platform.architecture()[0] == "32bit":
+        #    self.game_info['exe'] = line[str.index(line, ":") + 1:].strip()
+        # if line.startswith("exec_x64") and platform.architecture()[0] == "64bit":
+        #    self.exe = line[str.index(line, ":") + 1:].strip()
+        # if line.startswith('runner'):
+        #    self.game_info['runner'] = line[str.index(line, ":") + 1:]
+        #=======================================================================
 
     def writeConfig(self):
-        config_filename = self.lutrisConfigDir + "games/" + self.gameName + ".conf"
+        config_filename = os.path.join(self.lutrisConfigDir, "games" , self.gameName + ".conf")
         config_file = open(config_filename, "w")
         config_file.write("[main]\n")
         config_file.write("path = " + self.gameDir + "\n")
-        if(self.exe != None):
-            config_file.write("exe = " + self.exe + "\n")
-        config_file.write("system = " + self.platform)
+        if('exe' in self.game_info):
+            config_file.write("exe = " + self.game_info['exe'] + "\n")
+        config_file.write("runner = " + self.game_info['runner'])
         config_file.close()
 
     def reporthook(self, arg1, bytes, totalSize):
@@ -62,14 +121,40 @@ class Installer():
         print self.totalDownloaded / totalSize * 100.0
 
     def download(self, url):
-        url = url.strip()
-        filename = url.split("/")[-1]
-        destFile = self.gamesDir + self.gameName + "/" + filename
+        logging.debug('Downloading ' + url)
+        destFile = os.path.join(self.gamesDir, self.gameName, url.split("/")[-1])
+        if os.path.exists(destFile):
+            return destFile
         if url.startswith("file://"):
             shutil.copy(url[7:], self.gameDir)
         else:
             urllib.urlretrieve(url, destFile, self.reporthook)
         return destFile
+
+    def check_md5(self, data):
+        print 'checking md5 for file ' + self.gamefiles[data['file']]
+        print 'expecting ' + data['value']
+        print "NOT IMPLEMENTED"
+        return True
+
+    def extract(self, data):
+        print 'extracting ' + data['file']
+        filename = self.gamefiles[data['file']]
+        print "NOT IMPLEMENTED"
+        extension = filename[filename.rfind(".") + 1:]
+
+        if extension == "zip":
+            self.unzip(filename)
+
+    def move(self, data):
+        src = data['src']
+        destination_alias = data['dst']
+        if destination_alias == 'gamedir':
+            dst = self.gameDir;
+        else:
+            dst = '/tmp'
+        print "Moving %s to %s" % (src, dst)
+        shutil.move(src, dst)
 
     def unzip(self, file):
         subprocess.Popen(["unzip", file])
@@ -85,14 +170,6 @@ class Installer():
         subprocess.Popen([file])
 
 
-class WebInstaller():
-    def __init__(self):
-        self.website_url = "http://localhost:8000"
-    def install_game(self, name):
-        self.read_installer_desc(name)
-
-    def read_installer(self, name):
-
-
 if __name__ == "__main__":
-    quakeInstall = LutrisInstaller("hexen")
+    quakeInstall = LutrisInstaller("quake")
+    quakeInstall.install()
